hi yes i got it wrong again. correct now, i think.
yes, technically, since this is for i386 which has a 32-bit size_t,
it should be perfectly fine. however, it was using 4 bytes more than
necessary and that bothered me.

+++ src/libraries/hal-i386/basic_display.c
@@ -2,7 +2,6 @@
 
 #include <kernel/hal_basic_display.h>
 #include <stdint.h>
-#include <stddef.h>
 
 const uint16_t *VIDEO_RAM = (const uint16_t*)0xB8000;
 
@@ -23,8 +22,8 @@ void hal_basic_display_print(const char *string)
     static uint8_t row = 0;
     static uint8_t col = 0;
 
-    size_t text_index;
-    size_t color_index;
+    uint16_t text_index;
+    uint16_t color_index;
 
     if (video == 0) {
         video = (char*)VIDEO_RAM;
